requirements.txt
# Core dependencies
nats-py==2.7.2
pymongo==4.4.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Security
PyJWT==2.8.0
Werkzeug==2.3.6

# Logging and monitoring
prometheus-client==0.19.0

# Async support
aiofiles==23.2.1